server/app/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
python-dotenv
Flask-JWT-Extended
Flask-Migrate
werkzeug